Move mesh model defines
