+++ samples/common/mesh_model_defines/robot_movement_cli.h
